+++ src/commands/data/modelEntity.module.ts
@@ -32,7 +32,7 @@ export class ModelEntityModule extends ModelEntity {
     }
   }
 
-  protected openModelEntity(cn: string, parentName: string, relationshipNames: string[]) {
+  protected openModelEntity(cn: string, parentName: string, relationshipNames: string[], libDependencies: string[]) {
     const parentObject = parentName || "MIOManagedObject";
     let referenceParent = "";
     if (parentName) {
@@ -42,8 +42,10 @@ export class ModelEntityModule extends ModelEntity {
     }
 
     const importRelationships = relationshipNames.map((name) => `import { ${name} } from "./${name}";`);
+    const importFromLib = libDependencies.map((name) => `import { ${name} } from "miojslibs-core";`);
 
     const appendableContent = [
+      ...importFromLib,
       referenceParent,
       ...importRelationships,
       "",